refactor: move ConnectionInfo to own file
